test: BoxContainer
